function/requirements.txt
azure-common==1.1.25
azure-servicebus==0.50.0
psycopg2-binary==2.8.6
sendgrid==6.3.0